requirements.txt

gunicorn
Flask
requests
pdfplumber
python-dotenv

